Refactor submariner_controller_test
Add testDriver struct and other simplifications.
